services/IBF-pipeline/pipeline/requirements.txt
affine==2.3.0
area==1.1.1
asn1crypto==0.24.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cftime==1.4.1
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cryptography==2.1.4
Fiona==1.8.18
Flask==1.1.2
geopandas==0.8.2
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==10.6.0
keyrings.alt==3.0
loggly-python-handler==1.0.1
lxml==4.6.2
mailchimp3==3.0.14
MarkupSafe==1.1.1
munch==2.5.0
netCDF4==1.5.5.1
numpy==1.19.5
pandas==1.1.5
psycopg2-binary==2.8.6
pycrypto==2.6.1
pycurl==7.43.0.6
pygobject==3.26.1
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==3.0.0.post1
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
pytz==2021.1
pyxdg==0.25
rasterio==1.2.0
rasterstats==0.15.0
requests==2.25.1
requests-futures==1.0.0
scipy==1.5.4
SecretStorage==2.3.1
Shapely==1.7.1
simplejson==3.17.2
six==1.11.0
snuggs==1.4.7
soupsieve==2.1
SQLAlchemy==1.3.23
twilio==6.51.1
unattended-upgrades==0.1
urllib3==1.26.3
Werkzeug==1.0.1
xarray==0.16.2